Add Replayable Trx Scripts
This patch updates Trx Scripts to be replayable. Replayable scripts can be submitted multiple times from any sender. Special care should be taken to ensure that replayable scripts are safe to be replayed (e.g. checking state via storage, etc), but a good replayable script can add a ton of value to an account, e.g. crafting a script that buys Eth if it's less than $1000, only on Tuesdays, for up to 3 ETH, etc. This patch will be updated with test cases to see how it functions.