Add Conditionals via DelegateMulticall
This patch simply adds a new Quark Script (which are just helpers that don't need to be in the core, but they are since it's nice to expect them to already be deployed). This one allows you to specify that certain code gets "delegatecalled" (as opposed to .call'd) and thus will run on your account. These can be used as conditionals (e.g. check the day of the week or make sure this only runs X times per day or whatever). We currently use `.delegatecall`, though it's possible that we'd want .call or .callstatic, -- open questions here! But I'll apply this going forward to allow complex interactions at the app layer.