Add Signature Sets
This patch adds signature sets, which are the same as standard signatures for EIP-712 signed transaction scripts, however the signers are combined (via overflowing addition) to derive a _new address_, thus acting as a group (imagine being Megazord as opposed to Tyrannosaurus Zord, the Mastodon Zord, Pterodactyl Zord, Triceratops Zord, and Sabretooth Tiger Zord). This allows stateless X of X (not X of Y) MPC control of a unique wallet, which allows for more protected accounts. Note: wallets derived with 1 of 1 will be identical to a direct wallet, but 2 of 2 or more will be unable to sign a direct transaction and will _only_ be able to submit transactions via an EIP-712 signature.

Overall, this seems like a fair balance on how to get MPC without any of the complexity of MPC nor state.

Note: the addition operator, as opposed to SHA3 operations are still under evaluation.