refactor: allow nested nonce locks and nested callbacks, and cleanly
rather than try to set the nonce after an operation executes, we can set
it in the yieldTarget during the QuarkStorageManager callback and before
the nonce is released. this allows nested nonce locks.

and with nested nonce locks, and by scoping callbacks to the current
nonce, it's trivial to also support nested callbacks safely and cleanly.