[WIP] Slimmed down State Manager
This is just an exploration of a slimmed down `QuarkStateManager` from 🧇's PR. Specifically, we try to reduce the scope of responsibility in state manager to only really manage nonces, "active nonces" and opaque "data". The flow is the same where a wallet claims an active nonce (during script execution), receives a callback from QuarkStateManager, and then delegates execution. The biggest changes are:

1. Instead of having a "yieldTarget"- the script is able to set storage for `callback.v1.quark` in StateManager. We assert here that that is not priveleged since it's still isolated to the current script and thus is nothing the script couldn't do on its own.
2. We simplify setting nonce by making it `setNonce()` with no parameters (using the `activeNonce` field) and make it optional (e.g. for replayable scripts that don't want to necessarily set it).
3. We change namespaced storage to be `data` and key it by `bytes32` key, to allow scripts to store any number of items.
4. We move a bit more logic to the wallet and try to make the `QuarkStateManager` slimmer than before.

Overall, I think these changes make the system a bit cleaner. Note: most tests are passing on this branch. I believe this approach (either way) prevents nested scripts from executing in the same wallet, but we can dig deeper into this path if we want.