Execute Quark operation with script address (#50)
Adds the ability in `QuarkWallet` to execute an operation using an address of an already deployed transaction script. This helps to cut down on the calldata size, which helps to save some gas.

**Gas Impact:**
As can be seen in the unit tests below, using the `scriptAddress` instead of `scriptSource` helps to save 3-20k of gas depending on the calldata size.

```
[PASS] testAtomicIncrementerWithScriptAddress() (gas: 67237)
[PASS] testAtomicIncrementerWithScriptSource() (gas: 73484)

[PASS] testAtomicMaxCounterScriptWithScriptAddress() (gas: 178657)
[PASS] testAtomicMaxCounterScriptWithScriptSource() (gas: 197249)

[PASS] testAtomicPingWithScriptAddress() (gas: 48473)
[PASS] testAtomicPingWithScriptSource() (gas: 51042)

[PASS] testQuarkOperationWithScriptAddressRevertsIfCallReverts() (gas: 64952)
[PASS] testQuarkOperationWithScriptSourceRevertsIfCallReverts() (gas: 67981)

[PASS] testQuarkOperationWithScriptAddressRevertsIfCodeNotFound() (gas: 63602)
[PASS] testQuarkOperationWithScriptSourceRevertsIfCodeNotFound() (gas: 66064)
```

Note: For the `QuarkWallet` tests that use `scriptSource`, I am pre-deploying the script so the deployment gas costs are no longer included in the gas profiling. This allows the `scriptAddress` and `scriptSource` tests to be compared fairly.