feat: executor: an address allowed to execute unsigned code
this enables a wallet to act as a "primary account" wallet, which is
able to then issue commands to "sub-account" wallets. This maps fairly
cleanly to the idea of a "primary account" (ie. Schwab login) that
manages many "sub-account"s (ie. a Schwab taxable brokerage account).

QuarkWalletFactory is updated to take an opinionated view of wallet
relationships: the "default" wallet (salt=0) is the executor for all
sub-wallets with salt!=0. However, QuarkWallet itself is indifferent:
any arbitrary address can be passed into the constructor for executor.

this pattern unlocks the ability to perform cross-wallet actions with a
single signature and a single call; in other words, a multicall across
sub-account wallets. For example: transfer WBTC from wallet A to wallet
B, then supply from wallet B to Compound.