Add Out Of Memory Test
This test looks to be failing with "MemoryLimitOOG," which seems like it's out of gas based on the memory limit, which seems about as close to the solution here as we're going to get. Since Forge isn't exactly the true EVM, there will always be minor discrepencies, but that's making direct reference to the memory limit.