Add reentrancy guard modifier to QuarkScript; add more tests for re-entrancy (#95)
This introduces a `nonReentrant` modifier to `QuarkScript` that can be used by other scripts. We also improve our test coverage of reentrancy attacks by constructing a case of a successful attack that can occur when callback is enabled and the transfer is not protected by a reentrancy guard. We previously did not have a happy path of the attack since we didn't have callbacks enabled.