EIP-712 Fix
This patch is to fix the EIP-712 struct hash. Per the [EIP-712 Spec of structHash](https://eips.ethereum.org/EIPS/eip-712#definition-of-hashstruct), "The dynamic values bytes and string are encoded as a keccak256 hash of their contents." This will cause proper EIP-712 signatures to mismatch, but not caught because our tests use the same logic as our contracts.