Make executeQuarkOperation and executeScript no longer payable (#105)
We decided that there are no real use-cases where `executeQuarkOperation` and `executeScript` needs to be payable. Though we don't believe that the issue described in `[M-2] Griefing QuarkWallet operations with small native ETH amounts` is actually a griefing vector, we revisited the decision to make these functions payable and agreed to revert that decision. This removes a bit of complexity from the wallet design.

It is important to note that the wallet itself can still receive native ETH through its `receive` and `fallback` functions. The `fallback` function is `payable` to allow callbacks to send ETH to the wallet.