[WIP] Proxy Direct
This patch is an attempt at making the proxy even more slimmed down (by ignoring the niceties of the Solidity `delegatecall` contract) but otherwise keeping things in Solidity. We also are working on compiler settings to remove metadata, etc, to make the contract even smaller. Note: for observation, I've also added the optimized yul output.