Script for recurring purchases and tests for replayable txns (#113)
We add a replayable script for recurring purchases ("buy X of an ERC20-token every Y interval up to Z total amount"). 

Since replayable txns for the same script address are allowed to be submitted with different calldata, the same Quark nonce can be used for multiple recurring purchase orders; we show this in our tests. There is also a `cancel()` function that the user can sign to cancel a nonce and, therefore, all the recurring purchase orders using that nonce.

The gas cost to execute this script for the first time is 200k, which is pretty reasonable, given that the script is not compiled with `via-ir` and not fully optimized. Subsequent calls will also be cheaper due to storage slots already being written to.