[WIP] Code Jar Improvements, Deploy with Code Jar, and Verification
This patch improves CodeJar to deploy a new contract `CodeJarStub` instead of building the EVM opcodes by hand. This is because we can use `return()` yul-opcode in constructors to build custom code (weird, right!?). From this, we also build a way of deploying contracts using CodeJar (that is: content-addressible contracts-- the deploy address is based on the actual opcodes of the contract you are deploying). Further, we work on `verify` functionality for these contracts, which works if (and right now, only if) the deployed contract doesn't use `const immutable`s. There are reasons this could break otherwise, but for simple contracts, it works.