Det, The Signing Proxy
This patch adds `det`, which is a proxy that always signs with the same exact signature. This allows for deterministic deployment addresses since it's a variety of content-addressable signatures. See the README.md in `det/` for more details.

Note: I can't get it to work right now, but it... should... work? Like it's accepted to the mempool _then ignored_ which might indicate that the signature is not, in fact, valid, but looks valid to EthersJS and Infura.