Add CodeJar Factory
This patch changes CodeJar to be deployed from a factory, which just... deploys it. But it uses `CREATE2` such that CodeJar is _guaranteed_ to be the same bytecode if it's the same deployment address. This removes the vast majority of "attack vectors" from depending on CodeJar to be at the same address on all chains.