Create a `onlyWallet` guard to prevent reentrancy with lower guarantees (#140)
SEE EDIT

This gas optimized reentrancy guard (`nonReentrantOptimized`) saves around ~5k+ gas compared to the approach of writing the re-entrancy guard to `StateManager`. We leave a note suggesting that it is not as safe as the version that writes to `StateManager`. 

Regardless, this guard should be safe for most scripts since the modifier uses a storage slot addressed at a hash with an unknown string preimage. The guard is safe to use as long as scripts don't write directly to this storage slot in the same execution context.

I originally intended to add the guard to `Ethcall.sol` as well, but decided not to due to the extra 15k gas overhead. I also believe the `Ethcall` use-case is very different and it should really only be used to call into known and safe contracts. A note of caution is added to `Ethcall` to call out that it is not reentrancy safe.

EDIT: After some discussion, we decided to use the `onlyWallet` guard, which is a cheaper reentrancy guard with weaker guarantees. It protects against typical reentrancy paths that involve a `call`, but cannot protect against "recursive reentrancy" involving `delegatecall`s. That being said, reentrancy attacks rarely, if ever, involve `delegatecall`s and there are loads more to worry about when a script uses `delegatecall`.