Prevent direct calls to scripts that use `delegatecall` (#142)
This change will cause scripts that use `delegatecall` (e.g. `Multicall`) to revert if called directly. This is done to prevent griefers from self-destructing the contract.

The approach taken in this PR uses an `initialize` function that writes the script contract's address to storage. Every call to the script will check that the `msg.sender != cachedAddress` to ensure the call is never in the context of the script contract.

In terms of gas, this approach seems to introduce a 2.5-3k gas overhead.

Note: `UniswapFlashLoan` and `UniswapFlashSwapExactOut` are not susceptible to the same griefing vector. This is because the `delegatecall` in these scripts can only be accessed in a codepath where `msg.sender == valid Uniswap pool`, meaning that a griefer would have to use `UniswapScript.run` as the entry point for the attack. However, the `run` function in these scripts makes an `allowCallback()` call, which reverts immediately if not executed in the context of a Quark wallet. I added [tests](https://github.com/compound-finance/quark/pull/142/commits/600371827cdfb00d03331b307b27c09c72e2773f) to verify this revert behavior.