Add Batch Sig Algorithm
This patch adds a batch signature algorithm, by allowing a signer to pass in a pre-pad and post-pad. When those exist, we verify `keccak(prepad • eip712_keccak • postpad)`. We posit that this allows us to sign two EIP-712 signatures with a single digital signature. Specifically, the user will sign `keccak(eip712_keccak_0, eip712_keccak_1)`. Then when we want to verify the first one, we pass in the quark operation with a prepad="" postpad=eip712_keccak_1. To verify the second signature, we pass in the second quark operation with prepad=eip712_keccak_0 and postpad="". We posit that this is safe and effective.

As a note: we do not run the double keccak if prepad="" postpad="" to allow standard EIP-712 signatures to be normal.