Add MultiQuarkOperations (#186)
A different flavor of Batch sigs (#175).

`MultiQuarkOperation`s allows a signer to sign a single digest to execute multiple Quark Operations on any number of wallets owned by the signer on any number of chains. The most practical application of this is cross-chain operations, such as "Bridge USDC from mainnet to Base -> Supply USDC to a lending protocol on Base". 

`MultiQuarkOperation` signatures conform to EIP-712. This means that it has its own typehash and domain separator. The domain separator is special in that it does not include `chainId` and `verifyingContract`. This is to enable a single `MultiQuarkOperation` to support operations spanning different wallets and chains. This is safe because each `opsDigest` in the `MultiQuarkOperation` is already scoped to a specific `chainId` and `verifyingContract`, since each of these are a EIP-712 digest for a `QuarkOperation`.