Add Quotecall
Quotecall is like Paycall, except it always sends a specific quote of USDC to `tx.origin` allowing for minor changes in gas price (i.e. via a `deltaBps` parameter). As the USDC is sent _before_ the sub-transaction is run, this allow things like "send max USDC" to properly work as expected. As the quote is fixed, this means that recurring or delayed transactions are very _unlikely_ to work with this script.