Remove option to swallow reverts in Paycall and Quotecall (#199)
The `propagateReverts` constructor arg was introduced recently to
`Paycall` and `Quotecall` to support swallowing of reverts in the two
scripts. This was done to allow submitters to safely submit
`Paycall/Quotecall` operations on networks without Flashbots (e.g. all
L2s). Otherwise, without Flashbots' no-revert guarantee for
transactions, submitters will eat the cost for any reverts in
`Paycall/Quotecall`.

However, this option to swallow reverts introduces a potential gas
griefing vector. When reverts are swallowed, the submitter will still
get paid the payment token, at the expense of the user that signed the
Quark Operation. This is normally fine and expected, since users should
be paying for failing transactions and submitters need to be compensated
for doing the submission. This becomes problematic when the reverting
transaction is a replayable Quark Operation. In this case, a malicious
submitter could repeatedly submit the reverting replayable transaction,
draining the Quark wallet of its payment token in the process.

In short, we remove the option to swallow reverts in `Paycall/Quotecall`
to prevent this case from happening. Submitters on networks without
Flashbots are encouraged to simulate transactions first before
submitting them on-chain.