Quark Nonce Manager [Renaming] (#207)
This patch mostly renames `QuarkStateManager` to `QuarkNonceManager`.
Additionally, we start to use the term `submissionToken`, since it can
cover first plays or replays both as "submissions". We choose the term
`FREE` for the 0-token and `EXHAUSTED` for non-replayable tokens.
Overall, this should provide a clearer and more consistent naming
throughout the changeset.