Nonce as Non-Replayable Submission Token (#209)
Previously in `QuarkNonceManager`, we used `submissionToken=EXHAUSTED`
for non-replayable submissions and `submissionToken=nonce` for the first
play of replayable submissions. Overall, this behavior was inconsistent
(though otherwise fine) and generally unobservable. With the addition of
`getActiveSubmissionToken`, this behavior became obervable and showed
its inconsistencies. This patch changes the defined behavior making it
where we submit `submissionToken=nonce` for the first play or a
single-use or replayable quark operation, but still retain the behavior
that `submissions[nonce]=EXHAUSTED` for non-replayables. This ends up as
a reduction in the overall code complexity.

The only true behavior this changes is that we must (and should) ban
`nonce=bytes32(0)` (and for the sake of consistency also
`nonce=bytes32(uint_max)`). This is because we would be submitting with
`submissionToken=FREE` for that nonce and overall it isn't worth the
risk of allowing that (since it wouldn't, but _could_ lead to unlimited
replays if other code isn't implemented correctly).

Overall, I think this change is a net benefit in reduction of code
complexity and better understandability from the end-user's perspective.