TStore for Transients (#214)
This patch bumps us up to 0.8.27 and cancun and starts using `tstore`
and `tload` for transient reads and writes. Note: we still clear values
mostly to ensure that post-nested calls the script will see 0 instead of
the nested's own value.

Note: I have not taken a stab at understanding `ALLOW_CALLBACKS` and how
that should function.