Nonce Cancel by New Op (#213)
This patch adds the ability to cancel an outstanding nonce by sending a
new operation. Previously, it was required that the wallet had access to
the nonce chain of the previous operation to cancel that operation. We
now allow the wallet to craft a brand new operation that simply calls
cancel on a different nonce chain. This could also be used in a
bulk-cancelation script. Generally the process is safe as the new code
(a) depends on `msg.sender` and a properly executed script, and (b) only
works in the direction of cancelation.