Multi Quark Replays (#212)
This patch fixes the function names of and tests replaying multi-quark
operations. By and large, nothing changed and this works out of the box
more or less for free, but we want to make sure we test the behavior to
ensure there's no gotchas and that storage is correct, etc. So far, so
good.