Switch to isReplayable (#205)
This patch fixes some items in regards to a user signing `isReplayable`.
Specifically, when a user signs an operation, they need to specify if
the quark operation is replayable or not. When not, the state manager
nonce _must be_ `bytes32(uint256(-1))`, but when it is replayable, the
first play _must be_ `op.nonce`. If the user doesn't sign
`isReplayable`, then we'd need to enforce that the first play is
`replayToken = op.nonce`, which is bad since it that script could (if a
bad nonce is chosen) end up replayable, which is an attack vector for a
semi-bad client decision. Thus, we add this check to the user's
signature and all things should be good in the world.

Also, we start to fix a bit of verbiage to make things a bit nicer.

Techncially, while it's missing about 200 new tests, this should be a
complete implementation.

Whitespace changes for fmt check

Fix two minor comments

Minor gas improvements