Remove isolated storage and temporarily disable replays (#203)
As part of our redesign for Quark v2, we decided to scrap isolated
storage to simplify the flows for executing a Quark operation and lower
the gas overhead.

By removing isolated storage, we can remove the callback
(`setActiveNonceAndCallback`) from the `QuarkStateManager` to the
`QuarkWallet`. We can also remove the codepaths related to
`activeNonceScript`, `nonceScriptAddress`, and `walletStorage`. This
leaves only the nonce tracking logic in the `QuarkStateManager`.

We are also temporarily disabling replay functionality in this PR.
Replays will be brought back in a follow-up PR that implements the new
replay token mechanism. I've commented any tests that have to do with
replays, leaving a TODO (`// TODO: Uncomment when replay tokens are
supported`) to revisit once they are supported again.

There are also some TODOs to switch from sstore/ssload to tstore/tload.
To use tstore/tload, we need to bump up our version of Solidity to
>0.8.24. I plan to do that in a separate PR.