Add Replayables with Token (#204)
This code leads out the core changes for replayables with token. This
was meant to be a WIP that just showed off the core features (it still
is), but to get it to compile, I ended up converting most of the current
tests to the new framework (e.g. mostly converting them to pick a
semi-random nonce versus `nextNonce`, and checking `getNonceToken` as
opposed to `getNextNonce`). Overall, the changes to the non-test code
are very straight-forward. Most notable should be the changes to
`QuarkStateManager` (adding new nonce and replay token code), and adding
a new function to QuarkWallet
`verifySigAndExecuteReplayableQuarkOperation`.

Note: I didn't spend much time working on the outer interface for this,
and I need to consider how this works best with multi-quark operations,
but I wanted to get the outline first so we could discuss. There are
also failing test cases and absent test cases, but again, more of a
discussion point than a final product here.

---------

Co-authored-by: kevincheng96 <kevincheng96@hotmail.com>