Quark v2 (#216)
This is a moderately-sized redesign of Quark v1. The main goal of this
redesign is to improve the security of replayable operations by making
them only submittable by submitters that have access to a secret
submission token. A secondary goal of this redesign is to simplify a
bunch of the flows to reduce gas overhead and code complexity.

Changes:
- Introduce new replay mechanism that requires a submitter to provide a
`submissionToken`. Previously, replayable operations were submittable by
any address
- Remove isolated storage in the `QuarkStateManager`, which allows us to
remove much of the complicated codepaths related to
`setActiveNonceAndCallback`, `activeNonceScript`, `nonceScriptAddress`,
and `walletStorage`
- Rename `QuarkStateManager` to `QuarkNonceManager` since its job is now
to only manage nonces
- Adjust `QuarkOperation` struct to have a `isReplayable` field
- Bump up Solidity version to 0.8.27 and EVM version to `cancun`
- Use `TSTORE/TLOAD` to save gas

---------

Co-authored-by: Geoff Hayes <hayesgm@gmail.com>