Switch from `callcode` to `delegatecall` (#222)
Now that we use `tstore/tload` for reentrancy guards, the `onlyWallet`
modifier is no longer needed. This is was pretty much the only remaining
reason for using `callcode` over `delegatecall` to execute a Quark
script. Since `callcode` is a deprecated opcode, it is better to switch
over to `delegatecall`.

The one side effect of this switch-over is that scripts have to use
`address(this)` instead of `msg.sender` to refer to the Quark wallet's
address. Surprisingly, all our existing core scripts were already using
`address(this)`, so the only changes we have to make are to the test
scripts.