Use bytes for signature instead of v,r,s (#223)
`QuarkWallets` currently take `v`, `r`, `s` as inputs for the ECDSA
signature. This limits the wallet to only support ECDSA signatures, even
if the `signer` of the wallet is a EIP-1271 smart contract that supports
other signature schemes (e.g. BLS, Passkey). To make the wallet more
flexible and be able to support other signature schemes, we change the
signature input to a generic `bytes` type.

Note: This only increases flexibility for Quark wallets that have an
EIP-1271 contract as a signer, where the EIP-1271 contract is
upgradeable. Quark wallets owned by EOA signers will still need to be
migrated over to a new wallet implementation in order to support other
types of signature schemes.

Gas impact: Seems like this change bumps up gas usage by 900-1k
(negligible). This is likely because the wallet now needs to take the
extra step of decoding `bytes signature` into `v,r,s`. The tests that
fail the gas snapshot are all ones that run `executeQuarkOperation`
multiple times, exceeding the gas diff tolerance.